backend/requirements.txt
fastapi[all]>=0.115.6
uvicorn[standard]>=0.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.20
sqlalchemy>=2.0.37
pydantic>=2.10.5
pydantic-settings>=2.7.1
python-dotenv>=1.0.1
openai>=1.12.0 